# Add AVX-512 support for small FC layers