--- a/src/nnue/layers/affine_transform.h
+++ b/src/nnue/layers/affine_transform.h
@@ -470,7 +470,14 @@ namespace Stockfish::Eval::NNUE::Layers {
     const OutputType* propagate(
         const InputType* input, OutputType* output) const {
 
-#if defined (USE_AVX2)
+#if defined (USE_AVX512)
+      using vec_t = __m512i;
+      #define vec_setzero _mm512_setzero_si512
+      #define vec_set_32 _mm512_set1_epi32
+      #define vec_add_dpbusd_32 Simd::m512_add_dpbusd_epi32
+      #define vec_add_dpbusd_32x2 Simd::m512_add_dpbusd_epi32x2
+      #define vec_hadd Simd::m512_hadd
+#elif defined (USE_AVX2)
       using vec_t = __m256i;
       #define vec_setzero _mm256_setzero_si256
       #define vec_set_32 _mm256_set1_epi32
@@ -489,12 +496,14 @@ namespace Stockfish::Eval::NNUE::Layers {
 #if defined (USE_SSSE3)
       const auto inputVector = reinterpret_cast<const vec_t*>(input);
 
-      static_assert(OutputDimensions % OutputSimdWidth == 0 || OutputDimensions == 1);
+      constexpr std::size_t accumulatorElementCount = sizeof(vec_t) / sizeof(OutputType);
 
-      if constexpr (OutputDimensions % OutputSimdWidth == 0)
+      static_assert(OutputDimensions % accumulatorElementCount == 0 || OutputDimensions == 1);
+
+      if constexpr (OutputDimensions % accumulatorElementCount == 0)
       {
         constexpr IndexType NumChunks = ceil_to_multiple<IndexType>(InputDimensions, 8) / 4;
-        constexpr IndexType NumRegs = OutputDimensions / OutputSimdWidth;
+        constexpr IndexType NumRegs = OutputDimensions / accumulatorElementCount;
 
         const auto input32 = reinterpret_cast<const std::int32_t*>(input);
         const vec_t* biasvec = reinterpret_cast<const vec_t*>(biases);