@maximmasiutin 2023-04-06
# Patch ("New") vs Master ("Old") ARCH=x86-64-vnni512

## Intel Xeon 8272CL (Threads=70)

### 3s+0.02s; Hash=1024 

```
cutechess-cli -engine name=New cmd=./stockfish-x86-64-vnni512-gcc-avx512-affine-transform -engine name=Old cmd=./stockfish-x86-64-vnni512-gcc-master -each proto=uci tc=inf/3+0.02 option.Threads=70 option.Hash=1024 -recover -openings file=UHO_XXL_0.90_1.19.epd format=epd order=random plies=6 -repeat 2 -games 2000 -concurrency 1 -ratinginterval 100 -sprt elo0=0 elo1=2 alpha=0.05 beta=0.05 -wait 1000 >match-Xeon-8272CL-70-vnni512-3s-002s-1024.txt


Score of New vs Old: 542 - 522 - 936  [0.505] 2000
...      New playing White: 535 - 6 - 459  [0.764] 1000
...      New playing Black: 7 - 516 - 477  [0.245] 1000
...      White vs Black: 1051 - 13 - 936  [0.759] 2000
Elo difference: 3.5 +/- 11.1, LOS: 73.0 %, DrawRatio: 46.8 %
SPRT: llr 0.154 (5.2%), lbound -2.94, ubound 2.94

Player: New
   "Draw by 3-fold repetition": 480
   "Draw by fifty moves rule": 157
   "Draw by insufficient mating material": 269
   "Draw by stalemate": 30
   "Loss: Black mates": 4
   "Loss: White loses on time": 2
   "Loss: White mates": 516
   "Win: Black mates": 6
   "Win: White loses on time": 1
   "Win: White mates": 535
Player: Old
   "Draw by 3-fold repetition": 480
   "Draw by fifty moves rule": 157
   "Draw by insufficient mating material": 269
   "Draw by stalemate": 30
   "Loss: Black mates": 6
   "Loss: White loses on time": 1
   "Loss: White mates": 535
   "Win: Black mates": 4
   "Win: White loses on time": 2
   "Win: White mates": 516
Finished match

```

Patch wins (542 - 522 - 936  [0.505] 2000)

### 2s+0.1s; Hash=1024

```
cutechess-cli -engine name=New cmd=./stockfish-x86-64-vnni512-gcc-avx512-affine-transform -engine name=Old cmd=./stockfish-x86-64-vnni512-gcc-master -each proto=uci tc=inf/2+0.1 option.Threads=70 option.Hash=1024 -recover -openings file=UHO_XXL_0.90_1.19.epd format=epd order=random plies=6 -repeat 2 -games 2000 -concurrency 1 -ratinginterval 100 -sprt elo0=0 elo1=2 alpha=0.05 beta=0.05 -wait 1000


Score of New vs Old: 558 - 544 - 898  [0.503] 2000
...      New playing White: 557 - 1 - 442  [0.778] 1000
...      New playing Black: 1 - 543 - 456  [0.229] 1000
...      White vs Black: 1100 - 2 - 898  [0.774] 2000
Elo difference: 2.4 +/- 11.3, LOS: 66.3 %, DrawRatio: 44.9 %
SPRT: llr 0.0861 (2.9%), lbound -2.94, ubound 2.94

Player: New
   "Draw by 3-fold repetition": 346
   "Draw by fifty moves rule": 165
   "Draw by insufficient mating material": 365
   "Draw by stalemate": 22
   "Loss: Black mates": 1
   "Loss: White mates": 543
   "Win: Black mates": 1
   "Win: White mates": 557
Player: Old
   "Draw by 3-fold repetition": 346
   "Draw by fifty moves rule": 165
   "Draw by insufficient mating material": 365
   "Draw by stalemate": 22
   "Loss: Black mates": 1
   "Loss: White mates": 557
   "Win: Black mates": 1
   "Win: White mates": 543
Finished match
```

Patch wins (558 - 544 - 898  [0.503] 2000)

### 30s+1s; Hash=8192

```
cutechess-cli -engine name=New cmd=./stockfish-x86-64-vnni512-gcc-avx512-affine-transform -engine name=Old cmd=./stockfish-x86-64-vnni512-gcc-master -each proto=uci tc=inf/30+1 option.Threads=70 option.Hash=8192 -recover -openings file=UHO_XXL_0.90_1.19.epd format=epd order=random plies=6 -repeat 2 -games 200 -concurrency 1 -ratinginterval 100 -sprt elo0=0 elo1=2 alpha=0.05 beta=0.05 -wait 1000


Score of New vs Old: 50 - 49 - 101  [0.502] 200
...      New playing White: 50 - 0 - 50  [0.750] 100
...      New playing Black: 0 - 49 - 51  [0.255] 100
...      White vs Black: 99 - 0 - 101  [0.748] 200
Elo difference: 1.7 +/- 33.9, LOS: 54.0 %, DrawRatio: 50.5 %
SPRT: llr 0.00494 (0.2%), lbound -2.94, ubound 2.94

Player: New
   "Draw by 3-fold repetition": 33
   "Draw by fifty moves rule": 26
   "Draw by insufficient mating material": 42
   "Loss: White mates": 49
   "Win: White mates": 50
Player: Old
   "Draw by 3-fold repetition": 33
   "Draw by fifty moves rule": 26
   "Draw by insufficient mating material": 42
   "Loss: White mates": 50
   "Win: White mates": 49
Finished match
```

Finished equal (50 - 49 - 101  [0.502] 200)


## AMD Ryzen 7 7700 Zen 4 (Threads=16)

### 30s+1s; Hash=1024

```
cutechess-cli -engine name=New cmd=./stockfish-x86-64-vnni512-gcc-avx512-affine-transform -engine name=Old cmd=./stockfish-x86-64-vnni512-gcc-master -each proto=uci tc=inf/30+1 option.Threads=16 option.Hash=1024 -recover -openings file=UHO_XXL_0.90_1.19.epd format=epd order=random plies=6 -repeat 2 -games 200 -concurrency 1 -ratinginterval 100 -sprt elo0=0 elo1=2 alpha=0.05 beta=0.05 -wait 1000


Score of New vs Old: 63 - 53 - 84  [0.525] 200
...      New playing White: 63 - 0 - 37  [0.815] 100
...      New playing Black: 0 - 53 - 47  [0.235] 100
...      White vs Black: 116 - 0 - 84  [0.790] 200
Elo difference: 17.4 +/- 36.8, LOS: 82.3 %, DrawRatio: 42.0 %
SPRT: llr 0.0937 (3.2%), lbound -2.94, ubound 2.94

Player: New
   "Draw by 3-fold repetition": 18
   "Draw by fifty moves rule": 22
   "Draw by insufficient mating material": 41
   "Draw by stalemate": 3
   "Loss: White mates": 53
   "Win: White mates": 63
Player: Old
   "Draw by 3-fold repetition": 18
   "Draw by fifty moves rule": 22
   "Draw by insufficient mating material": 41
   "Draw by stalemate": 3
   "Loss: White mates": 63
   "Win: White mates": 53
Finished match
```

Patch wins (63 - 53 - 84  [0.525] 200).