@Disservin 2023-04-06
You don’t have to spend that much effort on the cutechess games, they are not telling us anything other than it doesn’t crash.