@maximmasiutin 2023-04-06
Closed this one because opened a new one https://github.com/official-stockfish/Stockfish/pull/4502 which produces identical binaries but presents the code differently, and updates the comments on the affected source code. The changes in the source code are suggested by @Sopel97